% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_data.R
\name{balanced.train.and.test}
\alias{balanced.train.and.test}
\title{Get a balanced test and train dataset}
\usage{
balanced.train.and.test(..., train.perc = 0.9, join.all = F)
}
\arguments{
\item{...}{vectors of index (could be numeric or logical)}

\item{train.perc}{percentage of dataset to be training set}

\item{join.all}{join all index in the end in two vectors (train and test vectors)}
}
\value{
train and test index vectors (two lists if `join.all = FALSE`, two vectors otherwise)
}
\description{
Get a balanced test and train dataset
}
\examples{
set1 <- c(1,2,3,5,6,7,8,9,10,11,12,13,14,15,17,18,19,20)
set2 <- c(F,F,F,T,F,F,F,F,F,F,F,F,F,F,F,T,F,F,F,F)
balanced.train.and.test(set1, set2, train.perc = .9, join.all = F)
####
set1 <- c(T,T,T,T,T,T,T,T,F,T,T,T,T,T,T,T,T,T,F,T)
set2 <- !set1
balanced.train.and.test(set1, set2, train.perc = .9, join.all = T)
}

