% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.genes.R
\name{ccds.genes.internal}
\alias{ccds.genes.internal}
\title{Download coding genes from CCDS}
\usage{
ccds.genes.internal()
}
\value{
vector of coding genes
}
\description{
https://ftp.ncbi.nih.gov/pub/CCDS/current_human/CCDS.current.txt
}
\examples{
\donttest{
  loose.rock:::ccds.genes.internal()
}
}
