% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loose.rock-package.r
\docType{data}
\name{coding.genes.cache}
\alias{coding.genes.cache}
\title{Coding gene cache for a faster vignette build}
\format{An object of class \code{data.frame} with 23099 rows and 2 columns.}
\usage{
coding.genes.cache
}
\description{
Coding gene cache for a faster vignette build
}
\keyword{datasets}
