% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.genes.R
\name{coding.genes.ensembl}
\alias{coding.genes.ensembl}
\title{Ensembl coding genes, local function}
\usage{
coding.genes.ensembl(verbose = TRUE, useCache = TRUE)
}
\arguments{
\item{verbose}{show messages with number of genes retrieved}

\item{useCache}{Boolean indicating whether the results cache
should be used. Setting to FALSE will disable reading and
writing of the cache. This argument is likely to disappear
after the cache functionality has been tested more thoroughly.}
}
\value{
a list with coding genes, mart and whether biomaRt had
a problem, indicating that it shouldn't be used.
}
\description{
Ensembl coding genes, local function
}
\examples{
\donttest{
  res <- loose.rock:::coding.genes.ensembl(TRUE, TRUE)
  nrow(res)
  head(res)
}
}
