% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\docType{methods}
\name{draw.kaplan,list,matrix,data.frame-method}
\alias{draw.kaplan,list,matrix,data.frame-method}
\title{draw.kaplan}
\usage{
\S4method{draw.kaplan}{list,matrix,data.frame}(chosen.btas, xdata, ydata,
  probs = c(0.5, 0.5), filename = "SurvivalCurves", save.plot = FALSE,
  xlim = NULL, ylim = NULL, expand.yzero = FALSE,
  base.directory = file.path("output", "kaplan-meier"),
  legend.outside = FALSE)
}
\arguments{
\item{chosen.btas}{list.}

\item{xdata}{matrix.}

\item{ydata}{data.frame.}

\item{probs}{How to separate high and low risk patients 50\%-50\% is the default, but for top and bottom 40\% -> c(.4,.6)}

\item{filename}{Name of file if save.plot is TRUE}

\item{save.plot}{TRUE plots everything and save, FALSE only calculates p-valeu}

\item{xlim}{Optional argument to limit the x-axis view}

\item{ylim}{Optional argument to limit the y-axis view}

\item{expand.yzero}{expand to y = 0}

\item{base.directory}{Initial directory where to store files}

\item{legend.outside}{If TRUE legend will be outside plot, otherwise inside}
}
\value{
object with logrank test and kaplan-meier survival plot
}
\description{
draw.kaplan
}
