% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{draw.kaplan}
\alias{draw.kaplan}
\title{draw.kaplan}
\usage{
draw.kaplan(chosen.btas, xdata, ydata, probs = c(0.5, 0.5),
  filename = "SurvivalCurves", save.plot = FALSE, xlim = NULL,
  ylim = NULL, expand.yzero = FALSE, base.directory = file.path("output",
  "kaplan-meier"), legend.outside = FALSE)
}
\arguments{
\item{chosen.btas}{list of testing coefficients to calculate prognostic indexes, for example ``list(Age = some_vector)``}

\item{xdata}{n x m matrix with n observations and m variables}

\item{ydata}{Survival object}

\item{probs}{How to separate high and low risk patients 50\%-50\% is the default, but for top and bottom 40\% -> c(.4,.6)}

\item{filename}{Name of file if save.plot is TRUE}

\item{save.plot}{TRUE plots everything and save, FALSE only calculates p-valeu}

\item{xlim}{Optional argument to limit the x-axis view}

\item{ylim}{Optional argument to limit the y-axis view}

\item{expand.yzero}{expand to y = 0}

\item{base.directory}{Initial directory where to store files}

\item{legend.outside}{If TRUE legend will be outside plot, otherwise inside}
}
\value{
object with logrank test and kaplan-meier survival plot

A list with plot, p-value and kaplan-meier object
}
\description{
Mega function that draws multiple kaplan meyer survival curves (or just 1)
}
\examples{
data('ovarian', package = 'survival')
xdata <- ovarian[,c('age', 'resid.ds')]
ydata <- data.frame(time = ovarian$futime, status = ovarian$fustat)
draw.kaplan(c(age = 1, 0), xdata, ydata)
draw.kaplan(c(age = 1, 0.5), xdata, ydata)
draw.kaplan(c(age = 1), c(1,0,1,0,1,0), data.frame(time = runif(6), status = rbinom(6, 1, .5)))
}
