% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.genes.R
\name{getBM.internal}
\alias{getBM.internal}
\title{Internal call to biomaRt::getBM}
\usage{
getBM.internal(..., useCache = TRUE, verbose = FALSE)
}
\arguments{
\item{...}{Paramters for biomaRt::getBM}

\item{useCache}{biomaRt::getBM parameter to use local cache or not. We use it
to try again without cache in case of an error}

\item{verbose}{show more verbose}
}
\value{
result of the biomaRt::getBM call
}
\description{
Depending on R version (<4.0.0) then it needs to have a special call
}
\examples{
\donttest{
  mart <- loose.rock:::getHsapiensMart.internal()
  res <- loose.rock:::getBM.internal(
    attributes = c("ensembl_gene_id","external_gene_name"),
    filters    = 'biotype',
    values     = c('protein_coding'),
    mart       = mart,
  )
  nrow(res)
  head(res)
}
}
