% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.genes.R
\name{getHsapiensMart.internal}
\alias{getHsapiensMart.internal}
\title{Get hsapiens mart from biomaRt}
\usage{
getHsapiensMart.internal(
  verbose = FALSE,
  useCache = TRUE,
  domain = listEnsemblMirrors()
)
}
\arguments{
\item{verbose}{add extra information in messages}

\item{useCache}{use run.cache to speed up multiple calls}

\item{domain}{vector of possible domains to call biomaRt}
}
\value{
biomaRt hsapiens mart
}
\description{
Get hsapiens mart from biomaRt
}
\examples{
\donttest{
  loose.rock:::getHsapiensMart.internal()
  loose.rock:::getHsapiensMart.internal(verbose = TRUE, useCache = FALSE)
}
}
