% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.genes.R
\name{join.ensembl.and.ccds}
\alias{join.ensembl.and.ccds}
\title{Join genes from ensembl and ccds in a single table}
\usage{
join.ensembl.and.ccds(
  ensembl.genes,
  ccds.genes,
  mart,
  useCache = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{ensembl.genes}{protein coding genes from ensembl}

\item{ccds.genes}{protein coding genes from ccds}

\item{mart}{biomaRt dataset to use}

\item{useCache}{should biomart use cache}

\item{verbose}{show extra messages}
}
\value{
table with ensembl_gene_id and external_gene_name columns
}
\description{
[INTERNAL]
Finds the ensembl ids of genes from ccds
}
