% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.genes.R
\name{listEnsemblMirrors}
\alias{listEnsemblMirrors}
\title{Internal function to list mirrors}
\usage{
listEnsemblMirrors()
}
\value{
list of ensembl mirrors
}
\description{
Internal function to list mirrors
}
\examples{
loose.rock:::listEnsemblMirrors()
}
