% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{loose.rock.options}
\alias{loose.rock.options}
\title{Constants for 'loose.rock' package}
\usage{
loose.rock.options(..., simplify = FALSE, update = list())
}
\arguments{
\item{...}{TODO}

\item{simplify}{TODO}

\item{update}{pair list of update to options}
}
\description{
Log level constants and the logger options.
}
\details{
The logging configuration is managed by 'loose.rock.options', a function
generated by OptionsManager within 'futile.options'.
}
\seealso{
\code{futile.options}
}
