% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{my.plot.residuals}
\alias{my.plot.residuals}
\title{Plot multiple residuals}
\usage{
my.plot.residuals(my.residuals, prefix, my.ylim = NULL, my.xlim = NULL,
  filename = NULL, title = "")
}
\arguments{
\item{title}{}
}
\description{
Plot multiple residuals
}

