% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ggplot2.R
\name{my.save.ggplot}
\alias{my.save.ggplot}
\title{save ggplots}
\usage{
my.save.ggplot(filename, my.plot = last_plot(), base.directory,
  out.format = c("pdf", "png"), width = 6, height = 4,
  separate.directory = T)
}
\description{
save ggplots
}

