% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot.R
\name{my.save.plot}
\alias{my.save.plot}
\title{Function to save plots to multiple formats}
\usage{
my.save.plot(filename, base.directory, out.format = c("pdf", "png"),
  width = 10, height = 7, separate.directory = T)
}
\description{
Function to save plots to multiple formats
}

