% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{release_questions}
\alias{release_questions}
\title{Questions to ask when calling devtools::release()}
\usage{
release_questions()
}
\value{
vector of questions
}
\description{
This should be done when submitting to cran
}
