% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{methods}
\name{runCache,function-method}
\alias{runCache,function-method}
\title{Run function and save cache}
\usage{
\S4method{runCache}{`function`}(fun, ..., seed = NULL, base.dir = NULL,
  cache.prefix = "generic_cache", cache.digest = list(),
  show.message = NULL, force.recalc = FALSE, add.to.hash = NULL)
}
\arguments{
\item{fun}{function call name}

\item{...}{parameters for function call}

\item{seed}{when function call is random, this allows to set seed beforehand}

\item{base.dir}{directory where data is stored}

\item{cache.prefix}{prefix for file name to be generated from parameters (...)}

\item{cache.digest}{cache of the digest for one or more of the parameters}

\item{show.message}{show message that data is being retrieved from cache}

\item{force.recalc}{force the recalculation of the values}

\item{add.to.hash}{something to add to the filename generation}
}
\description{
DEPRECATED!!! use run.cache instead
}
\seealso{
run.cache
}
