% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{runCache}
\alias{runCache}
\title{Run function and save cache}
\usage{
runCache(fun, ..., base.dir = tempdir(), cache.prefix = "generic_cache",
  force.recalc = FALSE)
}
\arguments{
\item{force.recalc}{}
}
\description{
Run function and save cache
}
