% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{save.ggplot}
\alias{save.ggplot}
\title{Save ggplots to file in multiple formats}
\usage{
save.ggplot(filename, base.directory, my.plot = ggplot2::last_plot(),
  out.format = c("pdf", "png"), width = 6, height = 4,
  separate.directory = TRUE)
}
\arguments{
\item{filename}{name of file to save}

\item{base.directory}{base folder to save plot}

\item{my.plot}{plot to save}

\item{out.format}{format}

\item{width}{width of image}

\item{height}{height of image}

\item{separate.directory}{separate formats by directory}
}
\description{
Save ggplots to file in multiple formats
}
