% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{verissimo.options}
\alias{verissimo.options}
\title{Constants for 'verissimo'}
\usage{
verissimo.options(..., simplify = FALSE, update = list())
}
\arguments{
\item{...}{TODO}

\item{simplify}{TODO}

\item{update}{TODO}
}
\description{
Log level constants and the logger options.
}
\details{
The logging configuration is managed by 'verissimo.options', a function
generated by OptionsManager within 'futile.options'.
}
\seealso{
\code{futile.options}
}
\author{
Brian Lee Yung Rowe
}
\keyword{data}
