% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.genes.R
\name{coding.genes}
\alias{coding.genes}
\title{Retrive coding genes from known databases}
\usage{
coding.genes()
}
\value{
a table with gene information
}
\description{
Retrive coding genes from known databases
}
