% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.genes.R
\name{coding.genes}
\alias{coding.genes}
\title{Retrive coding genes from known databases}
\usage{
coding.genes(verbose = TRUE, useCache = TRUE, extra.verbose = FALSE)
}
\arguments{
\item{verbose}{show messages with number of genes retrieved}

\item{useCache}{Boolean indicating whether the results cache
should be used. Setting to FALSE will disable reading and
writing of the cache. This argument is likely to disappear
after the cache functionality has been tested more thoroughly.}

\item{extra.verbose}{This will make all function calls
verbose, which could be a lot of information.}
}
\value{
a table with gene information
}
\description{
It retrieves from NCBI and ensembl
}
\examples{
# This can take a few minutes depending on the connection
\donttest{
  res <- coding.genes()
  nrow(res)
  head(res)
}
}
