% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{draw.cov.matrix}
\alias{draw.cov.matrix}
\title{Plot covariance heatmap from matrix}
\usage{
draw.cov.matrix(my.matrix, fun = stats::cov, ...)
}
\arguments{
\item{my.matrix}{matrix to calculate the covariance}

\item{fun}{function to use}

\item{...}{arguments to pass to fun function}
}
\value{
a ggplot2
}
\description{
Plot covariance heatmap from matrix
}
\examples{
draw.cov.matrix(matrix(rnorm(100), ncol = 10))
draw.cov.matrix(gen.synth.xdata(10, 10, .2))
}
