% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{draw.kaplan}
\alias{draw.kaplan}
\title{draw.kaplan}
\usage{
draw.kaplan(chosen.btas, xdata, ydata, probs = c(0.5, 0.5),
  filename = "SurvivalCurves", save.plot = F, xlim = NULL, ylim = NULL,
  expand.yzero = F, base.directory = file.path("output", "kaplan-meier"))
}
\arguments{
\item{chosen.btas}{list of testing coefficients to calculate prognostic indexes}

\item{xdata}{n x m matrix with n observations and m variables}

\item{ydata}{Survival object}

\item{filename}{name of file to save the plot}

\item{save.plot}{TRUE plots everything and save, FALSE only calculates p-valeu}

\item{xlim}{Optional argument to limit the x-axis view}

\item{sep}{How to separate high and low risk patients 50\%-50\% is the default, but for top and bottom 40\% -> c(.4,.6)}
}
\value{
A list with plot, p-value and kaplan-meier object
}
\description{
Mega function that draws multiple kaplan meyer survival curves (or just 1)
}
\examples{

}

