% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ggplot2.R
\name{draw.kaplan}
\alias{draw.kaplan}
\title{draw.kaplan}
\usage{
draw.kaplan(filename = "kaplan", chosen.btas, xdata, ydata, sep = c(0.5,
  0.5), save.plot = T, xlim = NULL)
}
\arguments{
\item{filename}{name of file to save the plot}

\item{chosen.btas}{list of testing coefficients to calculate prognostic indexes}

\item{xdata}{n x m matrix with n observations and m variables}

\item{ydata}{Survival object}

\item{sep}{How to separate high and low risk patients 50\%-50\% is the default, but for top and bottom 40\% -> c(.4,.6)}

\item{save.plot}{TRUE plots everything and save, FALSE only calculates p-valeu}

\item{xlim}{Optional argument to limit the x-axis view}
}
\value{
A list with plot, p-value and kaplan-meier object
}
\description{
Mega function that draws multiple kaplan meyer survival curves (or just 1)
}
\examples{

}

