% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.genes.R
\name{getBM.internal}
\alias{getBM.internal}
\title{Internal call to biomaRt::getBM}
\usage{
getBM.internal(...)
}
\arguments{
\item{...}{see biomaRt::getBM as all parameters are the same}
}
\value{
result of the biomaRt::getBM call
}
\description{
Depending on R version (<4.0.0) then it needs to have a special call
}
\examples{
\donttest{
  mart <- loose.rock:::getHsapiensMart.internal()
  res <- loose.rock:::getBM.internal(
    attributes = c('ensembl_gene_id','external_gene_name'),
    filters    = 'biotype',
    values     = c('protein_coding'),
    mart       = mart,
    useCache   = FALSE
  )
  nrow(res)
  head(res)
}
}
\seealso{
biomaRt::getBM
}
