% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{my.plot.residuals}
\alias{my.plot.residuals}
\title{Plot multiple residuals}
\usage{
my.plot.residuals(my.residuals, prefix, my.ylim = NULL, my.xlim = NULL,
  filename = NULL, title = "", base.directory = file.path("output",
  "residuals"))
}
\arguments{
\item{my.residuals}{TODO}

\item{prefix}{TODO}

\item{my.ylim}{TODO}

\item{my.xlim}{TODO}

\item{filename}{TODO}

\item{title}{TODO}

\item{base.directory}{TODO}
}
\value{
a plot with residuals
}
\description{
Plot multiple residuals
}
