% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{my.save.ggplot}
\alias{my.save.ggplot}
\title{Save ggplots to file in multiple formats}
\usage{
my.save.ggplot(filename, my.plot = last_plot(), base.directory,
  out.format = c("pdf", "png"), width = 6, height = 4,
  separate.directory = T)
}
\arguments{
\item{filename}{name of file to save}

\item{my.plot}{plot to save}

\item{base.directory}{base folder to save plot}

\item{out.format}{format}

\item{width}{width of image}

\item{height}{height of image}

\item{separate.directory}{separate formats by directory}
}
\description{
Save ggplots to file in multiple formats
}
