% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{runCache}
\alias{runCache}
\title{Run function and save cache}
\usage{
runCache(fun, ..., seed = NULL, base.dir = tempdir.cache(),
  cache.prefix = "generic_cache", cache.digest = list(),
  show.message = TRUE, force.recalc = FALSE, add.to.hash = NULL)
}
\arguments{
\item{force.recalc}{}
}
\value{
the result of fun(...)
}
\description{
Run function and save cache
}
\examples{
runCache(c, 1, 2, 3, 4)
runCache(c, a=1, 2, c=3, 4) # should get result from cache
}
